requirements.txt
# python >= 3.8
PyYAML~=6.0
torch~=2.0.0
torchvision~=0.15.1
tqdm~=4.63.1
matplotlib~=3.7.1
pydicom~=2.3.1
lion_pytorch~=0.0.6